Use built-in WebServer functionality to serve files
Remove the function to serve files, as well as the extension to MIME
type translation, as this functionality is already part of the WebServer
class. Oddly enough, the WebServer example does not use it.

Rename index.html to index.htm, since that's what the WebServer code
redirects to when the path is just '/'. 2018 seems a bit late to still
be dealing with DOS file name restrictions.