Allow changing history interval from the browser
This adds functionality in the server to PUT a new value for the
historyInterval.

The HTML and JS are updated with a new tab that allows inspecting the
current value as well as saving a new one.

The configuration value is only stored in a variable in RAM so far.
Therefore, it will revert to the default on reboot.