Generate timestamps on the server side using SNTP
With the recently added ability to change the recording interval for the
history at runtime, the history entries are no longer guaranteed to be
evenly spaced. To fix this, assign each measurement a proper time stamp
and record it in the history.

In order to get wall clock time stamps, configure SNTP to obtain time
from the Internet. The time is seconds since 1970 UTC. Conversion to
local time happens in the browser.