Make reset button behaviour configurable
Add a configuration setting to determine whether the reset button should
function as a regular reset, or as an on/off button.

When used as an on/off button, the device will toggle between normal
working mode and deep sleep each time the button is pressed. Due to
contact bounce, this may not always work reliably.
Also note that when the reset button is configured as on/off, the device
will go into deep sleep mode when power is initially applied. This means
that an intermittent power failure will result in the device being
inactive, i.e. in deep sleep mode.