Add table with history plot to web interface
This adds a new tab with a chart. The data is the same as for the
history tab. Loading and displaying it is expensive, so activating
either the history or the plot tab will not trigger a history get,
unless there is no history loaded yet.
The reload buttons in the plot and the history tab will both trigger a
reload of the history data and thus cause updates in both tabs,
regardless of which one is being displayed.