Combine history graph and table in the same tab
Given that the history table and the graph both use the same data
source, it makes sense to combine them both in the same tab. An
additional button in the title bar allows switching between the table
and the graph.

For consistency, selecting the tab showing the history will once again
trigger a reload of the data, which is the same behaviour as for the
other tabs. Toggling between graph and table within the history tab will
not reload the data. Reloading can be done either using the reload
button, or by selecting the (already selected) tab again in the tab bar.